requirements.txt
appdirs==1.4.4
asgiref==3.2.10
CacheControl==0.12.6
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
config==0.5.0.post0
contextlib2==0.6.0.post1
defusedxml==0.6.0
distlib==0.3.1
distro==1.5.0
Django==3.1
django-allauth==0.42.0
django-cors-headers==3.5.0
django-debug-toolbar==3.1
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
filelock==3.0.12
html5lib==1.1
idna==2.10
msgpack==1.0.0
oauthlib==3.1.0
ordered-set==4.0.2
packaging==20.4
pep517==0.8.2
Pillow==8.0.0
progress==1.5
PyJWT==1.7.1
pyparsing==2.4.7
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
resolvelib==0.4.0
retrying==1.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
virtualenv==20.0.34
webencodings==0.5.1
